% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypkg2-package.R
\docType{package}
\name{mypkg2-package}
\alias{mypkg2}
\alias{mypkg2-package}
\title{mypkg2: TTS Hamilton Data}
\description{
At this point, I just want the package to work.
}
\author{
\strong{Maintainer}: Sadia Tasnim \email{tasnis1@mcmaster.ca}

}
\keyword{internal}
