% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypkg2-package.R
\docType{data}
\name{TTS_Hamilton}
\alias{TTS_Hamilton}
\title{TTS data for Hamilton.}
\format{
A data frame with 419 rows and 17 variables:
\describe{
\item{Ward}{Names of wards in Hamilton}
}
}
\source{
Transportation Tomorrow Survey (http://www.transportationtomorrow.on.ca/) Accessed: November 2, 2023
}
\usage{
data(TTS_Hamilton)
}
\description{
A dataset containing information about transportation characteristics of Hamilton.
}
\examples{
 data(TTS_Hamilton)
 length(TTS_Hamilton)
}
\keyword{datasets}
